Move CursorLoaders out to CursorLoadFactory
